The obstacles has acceleration dependent on what engine you choose. (This is before the full implementation of gearbox).

+++ App.h
@@ -32,7 +32,7 @@ class App: public GlutApp {
 
 	clock_t startTime;
 	clock_t endTime;
-	int elapsedTime;
+	float elapsedTime;
 	bool used;
 	bool used2;
 	int points;
